v1.5
Though I broke the code, turns out it was a memory issue. Fixed that and also changed some code so that this wouldn't happen in the future. This also allows for greater frequency ranges than before. Still trying to figure out UART, ADC, and GPIO Interupts. When those ar figured out then the code will be fully functional and ready to impliment, hopefully.